Move plugins to right dirs
